refactor: separate seed fixtures into separate files
fix: import properly
squash: fix merge conflict
