+++ backend/prisma/recommendedPlansData.ts
@@ -1,0 +1,136 @@
+// Recommended Plans
+//  We can figure out a better way to populate these later.
+const recommendedPlansData = [
+  {
+    ageInMonth: '1M',
+    numOfNaps: 5,
+    wakeWindow: 1
+  },
+  {
+    ageInMonth: '2M',
+    numOfNaps: 5,
+    wakeWindow: 1.25
+  },
+  {
+    ageInMonth: '3M',
+    numOfNaps: 5,
+    wakeWindow: 1.75
+  },
+  {
+    ageInMonth: '4M',
+    numOfNaps: 4,
+    wakeWindow: 2
+  },
+  {
+    ageInMonth: '5M',
+    numOfNaps: 4,
+    wakeWindow: 2.25
+  },
+  {
+    ageInMonth: '6M',
+    numOfNaps: 3,
+    wakeWindow: 2
+  },
+  {
+    ageInMonth: '7M',
+    numOfNaps: 3,
+    wakeWindow: 2.5
+  },
+  {
+    ageInMonth: '8M',
+    numOfNaps: 3,
+    wakeWindow: 3
+  },
+  {
+    ageInMonth: '9M',
+    numOfNaps: 2,
+    wakeWindow: 3
+  },
+  {
+    ageInMonth: '10M',
+    numOfNaps: 2,
+    wakeWindow: 3.25
+  },
+  {
+    ageInMonth: '11M',
+    numOfNaps: 2,
+    wakeWindow: 3.5
+  },
+  {
+    ageInMonth: '12M',
+    numOfNaps: 2,
+    wakeWindow: 3.75
+  },
+  {
+    ageInMonth: '13M',
+    numOfNaps: 2,
+    wakeWindow: 4
+  },
+  {
+    ageInMonth: '14M',
+    numOfNaps: 2,
+    wakeWindow: 4
+  },
+  {
+    ageInMonth: '15M',
+    numOfNaps: 2,
+    wakeWindow: 4
+  },
+  {
+    ageInMonth: '16M',
+    numOfNaps: 2,
+    wakeWindow: 4
+  },
+  {
+    ageInMonth: '17M',
+    numOfNaps: 2,
+    wakeWindow: 4
+  },
+  {
+    ageInMonth: '18M',
+    numOfNaps: 1,
+    wakeWindow: 5
+  },
+  {
+    ageInMonth: '19M',
+    numOfNaps: 1,
+    wakeWindow: 5.25
+  },
+  {
+    ageInMonth: '20M',
+    numOfNaps: 1,
+    wakeWindow: 5.5
+  },
+  {
+    ageInMonth: '21M',
+    numOfNaps: 1,
+    wakeWindow: 5.5
+  },
+  {
+    ageInMonth: '22M',
+    numOfNaps: 1,
+    wakeWindow: 5.5
+  },
+  {
+    ageInMonth: '23M',
+    numOfNaps: 1,
+    wakeWindow: 5.5
+  },
+  {
+    ageInMonth: '24M',
+    numOfNaps: 1,
+    wakeWindow: 5.75
+  },
+  {
+    ageInMonth: '25M',
+    numOfNaps: 1,
+    wakeWindow: 5.75
+  },
+  {
+    ageInMonth: '26M',
+    numOfNaps: 1,
+    wakeWindow: 6
+  }
+];
+
+export default recommendedPlansData;
