+++ backend/prisma/seed.ts
@@ -1,6 +1,7 @@
 import { PrismaClient } from '@prisma/client';
 import type { User, Baby, Event } from '@prisma/client';
-// import { v4 as uuidv4 } from 'uuid';
+import questionData from './questionData.js';
+import recommendedPlansData from './recommendedPlansData.js';
 
 const prisma = new PrismaClient();
 const TOTAL_CLIENTS: number = 30;
@@ -21,229 +22,6 @@ const possibleNames: string[] = [
   'Zach',
   'Joslin',
   'Curt'
-];
-
-// Questions
-//  We can figure out a better way to populate these later.
-const questionData = [
-  {
-    questionId: '1',
-    description: 'What is your first name?',
-    type: 'small-text',
-    onboarding_screen: 1
-  },
-  {
-    questionId: '2',
-    description: 'What is your last name?',
-    type: 'small-text',
-    onboarding_screen: 2
-  },
-  {
-    questionId: '3',
-    description: "What is your Baby's name?",
-    type: 'small-text',
-    onboarding_screen: 3
-  },
-  {
-    questionId: '4',
-    description: 'When was your Baby born?',
-    type: 'date',
-    onboarding_screen: 4
-  },
-  {
-    questionId: '5',
-    description: 'Does your child have any medical conditions?',
-    type: 'large-text',
-    onboarding_screen: 5
-  },
-  {
-    questionId: '6',
-    description:
-      "Briefly describe your child's sleep environment - Where do they sleep? Who do they sleep with? Do they have a routine? What is the environment like? ",
-    type: 'large-text',
-    onboarding_screen: 6
-  },
-  {
-    questionId: '7',
-    description:
-      "Please tell me briefly what your child's day looks like in terms of sleep and feeds: Example: 7am wakes up, 9am eats, 10am nap...",
-    type: 'large-text',
-    onboarding_screen: 7
-  },
-  {
-    questionId: '8',
-    description: 'How does your child fall asleep at night?',
-    type: 'large-text',
-    onboarding_screen: 8
-  },
-  {
-    questionId: '9',
-    description: 'How many times does your child wake up per night?',
-    type: 'number',
-    onboarding_screen: 9
-  },
-  {
-    questionId: '10',
-    description:
-      'Have you tried any methods to help your baby sleep better in the past? Tell me in detail about your experience.',
-    type: 'large-text',
-    onboarding_screen: 10
-  },
-  {
-    questionId: '11',
-    description: 'What sleep goals would you like to achieve?',
-    type: 'large-text',
-    onboarding_screen: 11
-  },
-  {
-    questionId: '12',
-    description:
-      'Have you started implementing my Baby/Toddler Sleep Learning program?',
-    type: 'yes-no',
-    onboarding_screen: 12
-  },
-  {
-    questionId: '13',
-    description:
-      "Please tell me in detail any additional information that can help me understand your child & family's sleep situation.",
-    type: 'large-text',
-    onboarding_screen: 13
-  }
-];
-
-// Recommended Plans
-//  We can figure out a better way to populate these later.
-const recommendedPlansData = [
-  {
-    ageInMonth: '1M',
-    numOfNaps: 5,
-    wakeWindow: 1
-  },
-  {
-    ageInMonth: '2M',
-    numOfNaps: 5,
-    wakeWindow: 1.25
-  },
-  {
-    ageInMonth: '3M',
-    numOfNaps: 5,
-    wakeWindow: 1.75
-  },
-  {
-    ageInMonth: '4M',
-    numOfNaps: 4,
-    wakeWindow: 2
-  },
-  {
-    ageInMonth: '5M',
-    numOfNaps: 4,
-    wakeWindow: 2.25
-  },
-  {
-    ageInMonth: '6M',
-    numOfNaps: 3,
-    wakeWindow: 2
-  },
-  {
-    ageInMonth: '7M',
-    numOfNaps: 3,
-    wakeWindow: 2.5
-  },
-  {
-    ageInMonth: '8M',
-    numOfNaps: 3,
-    wakeWindow: 3
-  },
-  {
-    ageInMonth: '9M',
-    numOfNaps: 2,
-    wakeWindow: 3
-  },
-  {
-    ageInMonth: '10M',
-    numOfNaps: 2,
-    wakeWindow: 3.25
-  },
-  {
-    ageInMonth: '11M',
-    numOfNaps: 2,
-    wakeWindow: 3.5
-  },
-  {
-    ageInMonth: '12M',
-    numOfNaps: 2,
-    wakeWindow: 3.75
-  },
-  {
-    ageInMonth: '13M',
-    numOfNaps: 2,
-    wakeWindow: 4
-  },
-  {
-    ageInMonth: '14M',
-    numOfNaps: 2,
-    wakeWindow: 4
-  },
-  {
-    ageInMonth: '15M',
-    numOfNaps: 2,
-    wakeWindow: 4
-  },
-  {
-    ageInMonth: '16M',
-    numOfNaps: 2,
-    wakeWindow: 4
-  },
-  {
-    ageInMonth: '17M',
-    numOfNaps: 2,
-    wakeWindow: 4
-  },
-  {
-    ageInMonth: '18M',
-    numOfNaps: 1,
-    wakeWindow: 5
-  },
-  {
-    ageInMonth: '19M',
-    numOfNaps: 1,
-    wakeWindow: 5.25
-  },
-  {
-    ageInMonth: '20M',
-    numOfNaps: 1,
-    wakeWindow: 5.5
-  },
-  {
-    ageInMonth: '21M',
-    numOfNaps: 1,
-    wakeWindow: 5.5
-  },
-  {
-    ageInMonth: '22M',
-    numOfNaps: 1,
-    wakeWindow: 5.5
-  },
-  {
-    ageInMonth: '23M',
-    numOfNaps: 1,
-    wakeWindow: 5.5
-  },
-  {
-    ageInMonth: '24M',
-    numOfNaps: 1,
-    wakeWindow: 5.75
-  },
-  {
-    ageInMonth: '25M',
-    numOfNaps: 1,
-    wakeWindow: 5.75
-  },
-  {
-    ageInMonth: '26M',
-    numOfNaps: 1,
-    wakeWindow: 6
-  }
 ];
 
 // Events
